# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mammoth API
  version: 0.1.0
  description: |
    Mammoth is a self-contained bare-metal provisioning engine.
    Send in an out-of-band address and a credential, get back a machine that runs.

    Conventions:
    - Errors follow RFC 9457 (`application/problem+json`) with machine-readable `code` and `retryable` extensions.
    - All creating POSTs accept an `Idempotency-Key` header; replays within the window return the original result.
    - List endpoints use cursor pagination (`page_size` + `cursor`, response `next_cursor`).
    - Every response carries `X-Request-Id`; the server contract version is returned in `X-Mammoth-API-Version`.
  license:
    name: Apache-2.0
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/jobs/{id}/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/jobId'
      - $ref: '#/components/parameters/taskId'
    get:
      operationId: getJobTask
      tags: [jobs]
      summary: Get a task with its stage details
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/jobs/{id}/tasks/{taskId}/logs:
    parameters:
      - $ref: '#/components/parameters/jobId'
      - $ref: '#/components/parameters/taskId'
    get:
      operationId: listJobTaskLogs
      tags: [jobs]
      summary: Task execution logs
      description: |
        Structured log lines emitted while the runner executed this task,
        persisted at write time (docs/02-architecture.md §5.2 log dual-write).
        Complements events — facts and state transitions — with the execution
        trail: what the runner did and observed, stage by stage. Cursor
        paginated over the monotonic log id (docs/03-api.md §1).
      parameters:
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Task log page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskLogList'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/jobs/{id}/tasks/{taskId}/logs/stream:
    parameters:
      - $ref: '#/components/parameters/jobId'
      - $ref: '#/components/parameters/taskId'
    get:
      operationId: streamJobTaskLogs
      tags: [jobs]
      summary: Stream task execution logs (SSE)
      description: |
        Server-sent events over the task's persisted log lines — the same
        records as listJobTaskLogs, tailed. The stream first replays the
        backlog after the resume point in id order, then pushes new lines as
        they land. The watermark is the monotonic log id: reconnect with
        Last-Event-ID to resume exactly where the connection dropped (it
        wins over the `after` query parameter). Once the task reaches a
        terminal state (succeeded/failed/canceled — interrupted stays open,
        a retry resumes the same task) and the backlog has drained, the
        server emits a final `eos` event and closes the stream.
      parameters:
        - name: after
          in: query
          description: Replay logs after this log id.
          schema:
            type: string
        - name: Last-Event-ID
          in: header
          schema:
            type: string
      responses:
        '200':
          description: SSE stream (`text/event-stream`); `log` frames, then a terminal `eos` frame.
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/TaskLog'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/jobs/{id}/events:
    parameters:
      - $ref: '#/components/parameters/jobId'
    get:
      operationId: streamJobEvents
      tags: [jobs]
      summary: Stream a job's events (SSE)
      description: |
        Server-sent events for a job: stage/state changes and machine facts.
        Watermark by event id; reconnects resume from the Last-Event-ID
        header. Bearer auth applies (integrators stream with a token).
      parameters:
        - name: Last-Event-ID
          in: header
          description: Resume the stream after this event id.
          schema:
            type: string
      responses:
        '200':
          description: SSE stream (`text/event-stream`).
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/events:
    get:
      operationId: listEvents
      tags: [jobs]
      summary: Query events (audit and SSE backlog)
      description: |
        Cursor-paginated event listing. Audit queries filter by
        resource/type; the SSE global stream is the same surface streamed
        via `streamEvents`.
      parameters:
        - name: resource_type
          in: query
          description: "`pending` covers zero-registration sightings; `webhook` webhook lifecycle."
          schema:
            type: string
            enum: [job, task, machine, credential, pending, webhook]
        - name: resource_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          description: Exact event type filter (e.g. machine.discovered).
          schema:
            type: string
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Event page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
  /api/v1/events/stream:
    get:
      operationId: streamEvents
      tags: [jobs]
      summary: Stream all events (SSE)
      parameters:
        - name: resource
          in: query
          description: Filter to one resource id (e.g. job_x9k2).
          schema:
            type: string
        - name: Last-Event-ID
          in: header
          schema:
            type: string
      responses:
        '200':
          description: SSE stream (`text/event-stream`).
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/Event'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Static API token managed by the deployment (first version); an external IdP adapter point is reserved.
...
  parameters:
...
    pageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    cursor:
      name: cursor
      in: query
      description: Opaque continuation token from a previous page (`next_cursor`).
      schema:
        type: string
...
    jobId:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/JobId'
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/TaskId'
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    MachineId:
      type: string
      pattern: '^mch_[0-9a-z]+$'
    JobId:
      type: string
      pattern: '^job_[0-9a-z]+$'
    TaskId:
      type: string
      pattern: '^tsk_[0-9a-z]+$'

    # ── meta ───────────────────────────────────────────────────────────────
...
    Problem:
      type: object
      description: RFC 9457 problem details with Mammoth extensions.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Machine-readable error code (`<DOMAIN>_<PHENOMENON>`).
          example: BMC_UNREACHABLE
        retryable:
          type: boolean
          description: Whether the same request may succeed on retry.
    ErrorInfo:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 UTC.

    # ── credentials ────────────────────────────────────────────────────────
...
    TaskState:
      type: string
      enum: [pending, running, succeeded, failed, canceled, interrupted]
    StageState:
      type: string
      enum: [pending, running, succeeded, failed, skipped, canceled]
...
    Stage:
      type: object
      required: [name, state]
      properties:
        name:
          type: string
        state:
          $ref: '#/components/schemas/StageState'
        attempt:
          type: integer
        started_at:
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
        finished_at:
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
        duration_ms:
          type: integer
          nullable: true
    Task:
      type: object
      required: [id, job_id, machine_id, state, attempt, stages, created_at]
      properties:
        id:
          $ref: '#/components/schemas/TaskId'
        job_id:
          $ref: '#/components/schemas/JobId'
        machine_id:
          $ref: '#/components/schemas/MachineId'
        state:
          $ref: '#/components/schemas/TaskState'
        attempt:
          type: integer
        flow_name:
          type: string
        answer_url:
          type: string
          nullable: true
          description: |
            Machine-facing URL of the rendered answer file (install flows).
            Exposed for audit and debugging; the task token in the path is
            the machine credential (docs/06-install-pipeline.md §2.2).
        stages:
          type: array
          items:
            $ref: '#/components/schemas/Stage'
        error:
          allOf:
            - $ref: '#/components/schemas/ErrorInfo'
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        finished_at:
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
...
    Event:
      type: object
      required: [id, resource_type, resource_id, type, ts]
      properties:
        id:
          type: integer
          format: int64
          description: Monotonic watermark (SSE Last-Event-ID).
        resource_type:
          type: string
        resource_id:
          type: string
        type:
          type: string
          example: task.stage_changed
        payload:
          additionalProperties: true
          nullable: true
        ts:
          $ref: '#/components/schemas/Timestamp'
    EventList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        next_cursor:
          type: string
          nullable: true
...
    TaskLog:
      type: object
      description: One persisted execution log line (docs/08-data-model.md §2).
      required: [id, level, message, ts]
      properties:
        id:
          type: integer
          format: int64
          description: Monotonic within the task; cursor for pagination.
        level:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR]
        stage:
          type: string
          nullable: true
        message:
          type: string
        attrs:
          type: object
          description: Structured log fields beyond stage (machine_id, job_id, code, …).
          additionalProperties: true
          nullable: true
        ts:
          $ref: '#/components/schemas/Timestamp'
    TaskLogList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskLog'
        next_cursor:
          type: string
          nullable: true
...